Move  src/controllers/activegate/internal/events/statefulset_event.go -> src/controllers/dynakube/activegate/statefulset/event.go
