+++ src/controllers/dynakube/activegate/statefulset/event.go
@@ -1,4 +1,4 @@
-package events
+package statefulset
 
 import v1 "k8s.io/api/apps/v1"
